Module megatiny_hal::attiny412pac

Peripheral access API for ATTINY412 microcontrollers (generated using svd2rust v0.24.1 ( ))

You can find an overview of the generated API here.

API features to be included in the next svd2rust release can be generated by cloning the svd2rust repository, checking out the above commit, and running cargo doc --open.

Modules

ac0

Analog Comparator

adc0

Analog to Digital Converter

bod

Bod interface

ccl

Configurable Custom Logic

clkctrl

Clock controller

cpu

CPU

cpuint

Interrupt Controller

crcscan

CRCSCAN

dac0

Digital to Analog Converter

evsys

Event System

fuse

Fuses

generic

Common register and bit access and modify traits

gpio

General Purpose IO

lockbit

Lockbit

nvmctrl

Non-volatile Memory Controller

porta

I/O Ports

portmux

Port Multiplexer

rstctrl

Reset controller

rtc

Real-Time Counter

sigrow

Signature row

slpctrl

Sleep Controller

spi0

Serial Peripheral Interface

syscfg

System Configuration Registers

tcb0

16-bit Timer Type B

tcd0

Timer Counter D

twi0

Two-Wire Interface

usart0

Universal Synchronous and Asynchronous Receiver and Transmitter

userrow

User Row

vporta

Virtual Ports

vportb

Virtual Ports

vportc

Virtual Ports

vref

Voltage reference

wdt

Watch-Dog Timer

Structs

AC0

Analog Comparator

ADC0

Analog to Digital Converter

BOD

Bod interface

CCL

Configurable Custom Logic

CLKCTRL

Clock controller

CPU

CPU

CPUINT

Interrupt Controller

CRCSCAN

CRCSCAN

DAC0

Digital to Analog Converter

EVSYS

Event System

FUSE

Fuses

GPIO

General Purpose IO

LOCKBIT

Lockbit

NVMCTRL

Non-volatile Memory Controller

PORTA

I/O Ports

PORTMUX

Port Multiplexer

Peripherals

All the peripherals

RSTCTRL

Reset controller

RTC

Real-Time Counter

SIGROW

Signature row

SLPCTRL

Sleep Controller

SPI0

Serial Peripheral Interface

SYSCFG

System Configuration Registers

TCB0

16-bit Timer Type B

TCD0

Timer Counter D

TWI0

Two-Wire Interface

USART0

Universal Synchronous and Asynchronous Receiver and Transmitter

USERROW

User Row

VPORTA

Virtual Ports

VPORTB

Virtual Ports

VPORTC

Virtual Ports

VREF

Voltage reference

WDT

Watch-Dog Timer

Enums

Interrupt

Enumeration of all the interrupts.

Constants

NVIC_PRIO_BITS

Number available in the NVIC for configuring priority