Type Definition megatiny_hal::attiny412pac::ac0::muxctrla::MUXPOS_R

pub type MUXPOS_R = FieldReader<u8, MUXPOS_A>;

Field MUXPOS reader - Positive Input MUX Selection

Implementations

impl MUXPOS_R

pub fn variant(&self) -> Option<MUXPOS_A>

Get enumerated values variant

pub fn is_pin0(&self) -> bool

Checks if the value of the field is PIN0

pub fn is_pin1(&self) -> bool

Checks if the value of the field is PIN1