Type Definition megatiny_hal::attiny412pac::adc0::ctrlc::PRESC_R

pub type PRESC_R = FieldReader<u8, PRESC_A>;

Field PRESC reader - Clock Pre-scaler

Implementations

impl PRESC_R

pub fn variant(&self) -> PRESC_A

Get enumerated values variant

pub fn is_div2(&self) -> bool

Checks if the value of the field is DIV2

pub fn is_div4(&self) -> bool

Checks if the value of the field is DIV4

pub fn is_div8(&self) -> bool

Checks if the value of the field is DIV8

pub fn is_div16(&self) -> bool

Checks if the value of the field is DIV16

pub fn is_div32(&self) -> bool

Checks if the value of the field is DIV32

pub fn is_div64(&self) -> bool

Checks if the value of the field is DIV64

pub fn is_div128(&self) -> bool

Checks if the value of the field is DIV128

pub fn is_div256(&self) -> bool

Checks if the value of the field is DIV256