Type Definition megatiny_hal::attiny412pac::adc0::muxpos::MUXPOS_R

pub type MUXPOS_R = FieldReader<u8, MUXPOS_A>;

Field MUXPOS reader - Analog Channel Selection Bits

Implementations

impl MUXPOS_R

pub fn variant(&self) -> Option<MUXPOS_A>

Get enumerated values variant

pub fn is_ain0(&self) -> bool

Checks if the value of the field is AIN0

pub fn is_ain1(&self) -> bool

Checks if the value of the field is AIN1

pub fn is_ain2(&self) -> bool

Checks if the value of the field is AIN2

pub fn is_ain3(&self) -> bool

Checks if the value of the field is AIN3

pub fn is_ain4(&self) -> bool

Checks if the value of the field is AIN4

pub fn is_ain5(&self) -> bool

Checks if the value of the field is AIN5

pub fn is_ain6(&self) -> bool

Checks if the value of the field is AIN6

pub fn is_ain7(&self) -> bool

Checks if the value of the field is AIN7

pub fn is_ain8(&self) -> bool

Checks if the value of the field is AIN8

pub fn is_ain9(&self) -> bool

Checks if the value of the field is AIN9

pub fn is_ain10(&self) -> bool

Checks if the value of the field is AIN10

pub fn is_ain11(&self) -> bool

Checks if the value of the field is AIN11

pub fn is_dac0(&self) -> bool

Checks if the value of the field is DAC0

pub fn is_intref(&self) -> bool

Checks if the value of the field is INTREF

pub fn is_tempsense(&self) -> bool

Checks if the value of the field is TEMPSENSE

pub fn is_gnd(&self) -> bool

Checks if the value of the field is GND