Type Definition megatiny_hal::attiny412pac::ccl::lut1ctrla::FILTSEL_R

pub type FILTSEL_R = FieldReader<u8, FILTSEL_A>;

Field FILTSEL reader - Filter Selection

Implementations

impl FILTSEL_R

pub fn variant(&self) -> Option<FILTSEL_A>

Get enumerated values variant

pub fn is_disable(&self) -> bool

Checks if the value of the field is DISABLE

pub fn is_synch(&self) -> bool

Checks if the value of the field is SYNCH

pub fn is_filter(&self) -> bool

Checks if the value of the field is FILTER