Type Definition megatiny_hal::attiny412pac::ccl::lut1ctrlc::INSEL2_R

pub type INSEL2_R = FieldReader<u8, INSEL2_A>;

Field INSEL2 reader - LUT Input 2 Source Selection

Implementations

impl INSEL2_R

pub fn variant(&self) -> Option<INSEL2_A>

Get enumerated values variant

pub fn is_mask(&self) -> bool

Checks if the value of the field is MASK

pub fn is_feedback(&self) -> bool

Checks if the value of the field is FEEDBACK

pub fn is_link(&self) -> bool

Checks if the value of the field is LINK

pub fn is_event0(&self) -> bool

Checks if the value of the field is EVENT0

pub fn is_event1(&self) -> bool

Checks if the value of the field is EVENT1

pub fn is_io(&self) -> bool

Checks if the value of the field is IO

pub fn is_ac0(&self) -> bool

Checks if the value of the field is AC0

pub fn is_tcb0(&self) -> bool

Checks if the value of the field is TCB0

pub fn is_tca0(&self) -> bool

Checks if the value of the field is TCA0

pub fn is_tcd0(&self) -> bool

Checks if the value of the field is TCD0

pub fn is_spi0(&self) -> bool

Checks if the value of the field is SPI0