Type Definition megatiny_hal::attiny412pac::evsys::asyncch1::ASYNCCH1_R

pub type ASYNCCH1_R = FieldReader<u8, ASYNCCH1_A>;

Field ASYNCCH1 reader - Asynchronous Channel 1 Generator Selection

Implementations

impl ASYNCCH1_R

pub fn variant(&self) -> Option<ASYNCCH1_A>

Get enumerated values variant

pub fn is_off(&self) -> bool

Checks if the value of the field is OFF

pub fn is_ccl_lut0(&self) -> bool

Checks if the value of the field is CCL_LUT0

pub fn is_ccl_lut1(&self) -> bool

Checks if the value of the field is CCL_LUT1

pub fn is_ac0_out(&self) -> bool

Checks if the value of the field is AC0_OUT

pub fn is_tcd0_cmpbclr(&self) -> bool

Checks if the value of the field is TCD0_CMPBCLR

pub fn is_tcd0_cmpaset(&self) -> bool

Checks if the value of the field is TCD0_CMPASET

pub fn is_tcd0_cmpbset(&self) -> bool

Checks if the value of the field is TCD0_CMPBSET

pub fn is_tcd0_progev(&self) -> bool

Checks if the value of the field is TCD0_PROGEV

pub fn is_rtc_ovf(&self) -> bool

Checks if the value of the field is RTC_OVF

pub fn is_rtc_cmp(&self) -> bool

Checks if the value of the field is RTC_CMP

pub fn is_portb_pin0(&self) -> bool

Checks if the value of the field is PORTB_PIN0

pub fn is_portb_pin1(&self) -> bool

Checks if the value of the field is PORTB_PIN1

pub fn is_portb_pin2(&self) -> bool

Checks if the value of the field is PORTB_PIN2

pub fn is_portb_pin3(&self) -> bool

Checks if the value of the field is PORTB_PIN3

pub fn is_portb_pin4(&self) -> bool

Checks if the value of the field is PORTB_PIN4

pub fn is_portb_pin5(&self) -> bool

Checks if the value of the field is PORTB_PIN5

pub fn is_portb_pin6(&self) -> bool

Checks if the value of the field is PORTB_PIN6

pub fn is_portb_pin7(&self) -> bool

Checks if the value of the field is PORTB_PIN7