Type Definition megatiny_hal::attiny412pac::evsys::asyncch3::ASYNCCH3_R

pub type ASYNCCH3_R = FieldReader<u8, ASYNCCH3_A>;

Field ASYNCCH3 reader - Asynchronous Channel 3 Generator Selection

Implementations

impl ASYNCCH3_R

pub fn variant(&self) -> Option<ASYNCCH3_A>

Get enumerated values variant

pub fn is_off(&self) -> bool

Checks if the value of the field is OFF

pub fn is_ccl_lut0(&self) -> bool

Checks if the value of the field is CCL_LUT0

pub fn is_ccl_lut1(&self) -> bool

Checks if the value of the field is CCL_LUT1

pub fn is_ac0_out(&self) -> bool

Checks if the value of the field is AC0_OUT

pub fn is_tcd0_cmpbclr(&self) -> bool

Checks if the value of the field is TCD0_CMPBCLR

pub fn is_tcd0_cmpaset(&self) -> bool

Checks if the value of the field is TCD0_CMPASET

pub fn is_tcd0_cmpbset(&self) -> bool

Checks if the value of the field is TCD0_CMPBSET

pub fn is_tcd0_progev(&self) -> bool

Checks if the value of the field is TCD0_PROGEV

pub fn is_rtc_ovf(&self) -> bool

Checks if the value of the field is RTC_OVF

pub fn is_rtc_cmp(&self) -> bool

Checks if the value of the field is RTC_CMP

pub fn is_pit_div8192(&self) -> bool

Checks if the value of the field is PIT_DIV8192

pub fn is_pit_div4096(&self) -> bool

Checks if the value of the field is PIT_DIV4096

pub fn is_pit_div2048(&self) -> bool

Checks if the value of the field is PIT_DIV2048

pub fn is_pit_div1024(&self) -> bool

Checks if the value of the field is PIT_DIV1024

pub fn is_pit_div512(&self) -> bool

Checks if the value of the field is PIT_DIV512

pub fn is_pit_div256(&self) -> bool

Checks if the value of the field is PIT_DIV256

pub fn is_pit_div128(&self) -> bool

Checks if the value of the field is PIT_DIV128

pub fn is_pit_div64(&self) -> bool

Checks if the value of the field is PIT_DIV64