Type Definition megatiny_hal::attiny412pac::evsys::asyncuser6::ASYNCUSER6_R

pub type ASYNCUSER6_R = FieldReader<u8, ASYNCUSER6_A>;

Field ASYNCUSER6 reader - Asynchronous User Ch 6 Input Selection - TCD0 Event 0

Implementations

impl ASYNCUSER6_R

pub fn variant(&self) -> Option<ASYNCUSER6_A>

Get enumerated values variant

pub fn is_off(&self) -> bool

Checks if the value of the field is OFF

pub fn is_syncch0(&self) -> bool

Checks if the value of the field is SYNCCH0

pub fn is_syncch1(&self) -> bool

Checks if the value of the field is SYNCCH1

pub fn is_asyncch0(&self) -> bool

Checks if the value of the field is ASYNCCH0

pub fn is_asyncch1(&self) -> bool

Checks if the value of the field is ASYNCCH1

pub fn is_asyncch2(&self) -> bool

Checks if the value of the field is ASYNCCH2

pub fn is_asyncch3(&self) -> bool

Checks if the value of the field is ASYNCCH3