Type Definition megatiny_hal::attiny412pac::evsys::syncch0::SYNCCH0_R

pub type SYNCCH0_R = FieldReader<u8, SYNCCH0_A>;

Field SYNCCH0 reader - Synchronous Channel 0 Generator Selection

Implementations

impl SYNCCH0_R

pub fn variant(&self) -> Option<SYNCCH0_A>

Get enumerated values variant

pub fn is_off(&self) -> bool

Checks if the value of the field is OFF

pub fn is_tcb0(&self) -> bool

Checks if the value of the field is TCB0

pub fn is_tca0_ovf_lunf(&self) -> bool

Checks if the value of the field is TCA0_OVF_LUNF

pub fn is_tca0_hunf(&self) -> bool

Checks if the value of the field is TCA0_HUNF

pub fn is_tca0_cmp0(&self) -> bool

Checks if the value of the field is TCA0_CMP0

pub fn is_tca0_cmp1(&self) -> bool

Checks if the value of the field is TCA0_CMP1

pub fn is_tca0_cmp2(&self) -> bool

Checks if the value of the field is TCA0_CMP2

pub fn is_portc_pin0(&self) -> bool

Checks if the value of the field is PORTC_PIN0

pub fn is_portc_pin1(&self) -> bool

Checks if the value of the field is PORTC_PIN1

pub fn is_portc_pin2(&self) -> bool

Checks if the value of the field is PORTC_PIN2

pub fn is_portc_pin3(&self) -> bool

Checks if the value of the field is PORTC_PIN3

pub fn is_portc_pin4(&self) -> bool

Checks if the value of the field is PORTC_PIN4

pub fn is_portc_pin5(&self) -> bool

Checks if the value of the field is PORTC_PIN5

pub fn is_porta_pin0(&self) -> bool

Checks if the value of the field is PORTA_PIN0

pub fn is_porta_pin1(&self) -> bool

Checks if the value of the field is PORTA_PIN1

pub fn is_porta_pin2(&self) -> bool

Checks if the value of the field is PORTA_PIN2

pub fn is_porta_pin3(&self) -> bool

Checks if the value of the field is PORTA_PIN3

pub fn is_porta_pin4(&self) -> bool

Checks if the value of the field is PORTA_PIN4

pub fn is_porta_pin5(&self) -> bool

Checks if the value of the field is PORTA_PIN5

pub fn is_porta_pin6(&self) -> bool

Checks if the value of the field is PORTA_PIN6

pub fn is_porta_pin7(&self) -> bool

Checks if the value of the field is PORTA_PIN7