Type Definition megatiny_hal::attiny412pac::evsys::syncch1::SYNCCH1_R

pub type SYNCCH1_R = FieldReader<u8, SYNCCH1_A>;

Field SYNCCH1 reader - Synchronous Channel 1 Generator Selection

Implementations

impl SYNCCH1_R

pub fn variant(&self) -> Option<SYNCCH1_A>

Get enumerated values variant

pub fn is_off(&self) -> bool

Checks if the value of the field is OFF

pub fn is_tcb0(&self) -> bool

Checks if the value of the field is TCB0

pub fn is_tca0_ovf_lunf(&self) -> bool

Checks if the value of the field is TCA0_OVF_LUNF

pub fn is_tca0_hunf(&self) -> bool

Checks if the value of the field is TCA0_HUNF

pub fn is_tca0_cmp0(&self) -> bool

Checks if the value of the field is TCA0_CMP0

pub fn is_tca0_cmp1(&self) -> bool

Checks if the value of the field is TCA0_CMP1

pub fn is_tca0_cmp2(&self) -> bool

Checks if the value of the field is TCA0_CMP2

pub fn is_portb_pin0(&self) -> bool

Checks if the value of the field is PORTB_PIN0

pub fn is_portb_pin1(&self) -> bool

Checks if the value of the field is PORTB_PIN1

pub fn is_portb_pin2(&self) -> bool

Checks if the value of the field is PORTB_PIN2

pub fn is_portb_pin3(&self) -> bool

Checks if the value of the field is PORTB_PIN3

pub fn is_portb_pin4(&self) -> bool

Checks if the value of the field is PORTB_PIN4

pub fn is_portb_pin5(&self) -> bool

Checks if the value of the field is PORTB_PIN5

pub fn is_portb_pin6(&self) -> bool

Checks if the value of the field is PORTB_PIN6

pub fn is_portb_pin7(&self) -> bool

Checks if the value of the field is PORTB_PIN7