Type Definition megatiny_hal::attiny412pac::fuse::syscfg1::SUT_R

pub type SUT_R = FieldReader<u8, SUT_A>;

Field SUT reader - Startup Time

Implementations

impl SUT_R

pub fn variant(&self) -> SUT_A

Get enumerated values variant

pub fn is_0ms(&self) -> bool

Checks if the value of the field is _0MS

pub fn is_1ms(&self) -> bool

Checks if the value of the field is _1MS

pub fn is_2ms(&self) -> bool

Checks if the value of the field is _2MS

pub fn is_4ms(&self) -> bool

Checks if the value of the field is _4MS

pub fn is_8ms(&self) -> bool

Checks if the value of the field is _8MS

pub fn is_16ms(&self) -> bool

Checks if the value of the field is _16MS

pub fn is_32ms(&self) -> bool

Checks if the value of the field is _32MS

pub fn is_64ms(&self) -> bool

Checks if the value of the field is _64MS