Type Definition megatiny_hal::attiny412pac::nvmctrl::ctrla::CMD_R

pub type CMD_R = FieldReader<u8, CMD_A>;

Field CMD reader - Command

Implementations

impl CMD_R

pub fn variant(&self) -> CMD_A

Get enumerated values variant

pub fn is_none(&self) -> bool

Checks if the value of the field is NONE

pub fn is_pagewrite(&self) -> bool

Checks if the value of the field is PAGEWRITE

pub fn is_pageerase(&self) -> bool

Checks if the value of the field is PAGEERASE

pub fn is_pageerasewrite(&self) -> bool

Checks if the value of the field is PAGEERASEWRITE

pub fn is_pagebufclr(&self) -> bool

Checks if the value of the field is PAGEBUFCLR

pub fn is_chiperase(&self) -> bool

Checks if the value of the field is CHIPERASE

pub fn is_eeerase(&self) -> bool

Checks if the value of the field is EEERASE

pub fn is_fusewrite(&self) -> bool

Checks if the value of the field is FUSEWRITE