Type Definition megatiny_hal::attiny412pac::rtc::ctrla::PRESCALER_R

pub type PRESCALER_R = FieldReader<u8, PRESCALER_A>;

Field PRESCALER reader - Prescaling Factor

Implementations

impl PRESCALER_R

pub fn variant(&self) -> PRESCALER_A

Get enumerated values variant

pub fn is_div1(&self) -> bool

Checks if the value of the field is DIV1

pub fn is_div2(&self) -> bool

Checks if the value of the field is DIV2

pub fn is_div4(&self) -> bool

Checks if the value of the field is DIV4

pub fn is_div8(&self) -> bool

Checks if the value of the field is DIV8

pub fn is_div16(&self) -> bool

Checks if the value of the field is DIV16

pub fn is_div32(&self) -> bool

Checks if the value of the field is DIV32

pub fn is_div64(&self) -> bool

Checks if the value of the field is DIV64

pub fn is_div128(&self) -> bool

Checks if the value of the field is DIV128

pub fn is_div256(&self) -> bool

Checks if the value of the field is DIV256

pub fn is_div512(&self) -> bool

Checks if the value of the field is DIV512

pub fn is_div1024(&self) -> bool

Checks if the value of the field is DIV1024

pub fn is_div2048(&self) -> bool

Checks if the value of the field is DIV2048

pub fn is_div4096(&self) -> bool

Checks if the value of the field is DIV4096

pub fn is_div8192(&self) -> bool

Checks if the value of the field is DIV8192

pub fn is_div16384(&self) -> bool

Checks if the value of the field is DIV16384

pub fn is_div32768(&self) -> bool

Checks if the value of the field is DIV32768