Type Definition megatiny_hal::attiny412pac::rtc::pitctrla::PERIOD_R

pub type PERIOD_R = FieldReader<u8, PERIOD_A>;

Field PERIOD reader - Period

Implementations

impl PERIOD_R

pub fn variant(&self) -> Option<PERIOD_A>

Get enumerated values variant

pub fn is_off(&self) -> bool

Checks if the value of the field is OFF

pub fn is_cyc4(&self) -> bool

Checks if the value of the field is CYC4

pub fn is_cyc8(&self) -> bool

Checks if the value of the field is CYC8

pub fn is_cyc16(&self) -> bool

Checks if the value of the field is CYC16

pub fn is_cyc32(&self) -> bool

Checks if the value of the field is CYC32

pub fn is_cyc64(&self) -> bool

Checks if the value of the field is CYC64

pub fn is_cyc128(&self) -> bool

Checks if the value of the field is CYC128

pub fn is_cyc256(&self) -> bool

Checks if the value of the field is CYC256

pub fn is_cyc512(&self) -> bool

Checks if the value of the field is CYC512

pub fn is_cyc1024(&self) -> bool

Checks if the value of the field is CYC1024

pub fn is_cyc2048(&self) -> bool

Checks if the value of the field is CYC2048

pub fn is_cyc4096(&self) -> bool

Checks if the value of the field is CYC4096

pub fn is_cyc8192(&self) -> bool

Checks if the value of the field is CYC8192

pub fn is_cyc16384(&self) -> bool

Checks if the value of the field is CYC16384

pub fn is_cyc32768(&self) -> bool

Checks if the value of the field is CYC32768