Type Definition megatiny_hal::attiny412pac::tcb0::ctrlb::CNTMODE_R

pub type CNTMODE_R = FieldReader<u8, CNTMODE_A>;

Field CNTMODE reader - Timer Mode

Implementations

impl CNTMODE_R

pub fn variant(&self) -> CNTMODE_A

Get enumerated values variant

pub fn is_int(&self) -> bool

Checks if the value of the field is INT

pub fn is_timeout(&self) -> bool

Checks if the value of the field is TIMEOUT

pub fn is_capt(&self) -> bool

Checks if the value of the field is CAPT

pub fn is_frq(&self) -> bool

Checks if the value of the field is FRQ

pub fn is_pw(&self) -> bool

Checks if the value of the field is PW

pub fn is_frqpw(&self) -> bool

Checks if the value of the field is FRQPW

pub fn is_single(&self) -> bool

Checks if the value of the field is SINGLE

pub fn is_pwm8(&self) -> bool

Checks if the value of the field is PWM8