Type Definition megatiny_hal::attiny412pac::tcd0::evctrlb::ACTION_R

pub type ACTION_R = BitReader<ACTION_A>;

Field ACTION reader - event action

Implementations

impl ACTION_R

pub fn variant(&self) -> ACTION_A

Get enumerated values variant

pub fn is_fault(&self) -> bool

Checks if the value of the field is FAULT

pub fn is_capture(&self) -> bool

Checks if the value of the field is CAPTURE