Type Definition megatiny_hal::attiny412pac::tcd0::inputctrla::INPUTMODE_R

pub type INPUTMODE_R = FieldReader<u8, INPUTMODE_A>;

Field INPUTMODE reader - Input mode

Implementations

impl INPUTMODE_R

pub fn variant(&self) -> Option<INPUTMODE_A>

Get enumerated values variant

pub fn is_none(&self) -> bool

Checks if the value of the field is NONE

pub fn is_jmpwait(&self) -> bool

Checks if the value of the field is JMPWAIT

pub fn is_execwait(&self) -> bool

Checks if the value of the field is EXECWAIT

pub fn is_execfault(&self) -> bool

Checks if the value of the field is EXECFAULT

pub fn is_freq(&self) -> bool

Checks if the value of the field is FREQ

pub fn is_execdt(&self) -> bool

Checks if the value of the field is EXECDT

pub fn is_wait(&self) -> bool

Checks if the value of the field is WAIT

pub fn is_waitsw(&self) -> bool

Checks if the value of the field is WAITSW

pub fn is_edgetrig(&self) -> bool

Checks if the value of the field is EDGETRIG

pub fn is_edgetrigfreq(&self) -> bool

Checks if the value of the field is EDGETRIGFREQ

pub fn is_lvltrigfreq(&self) -> bool

Checks if the value of the field is LVLTRIGFREQ