Type Definition megatiny_hal::attiny412pac::tcd0::inputctrlb::INPUTMODE_W

pub type INPUTMODE_W<'a, const O: u8> = FieldWriter<'a, u8, INPUTCTRLB_SPEC, u8, INPUTMODE_A, 4, O>;

Field INPUTMODE writer - Input mode

Implementations

impl<'a, const O: u8> INPUTMODE_W<'a, O>

pub fn none(self) -> &'a mut W

Input has no actions

pub fn jmpwait(self) -> &'a mut W

Stop output, jump to opposite compare cycle and wait

pub fn execwait(self) -> &'a mut W

Stop output, execute opposite compare cycle and wait

pub fn execfault(self) -> &'a mut W

stop output, execute opposite compare cycle while fault active

pub fn freq(self) -> &'a mut W

Stop all outputs, maintain frequency

pub fn execdt(self) -> &'a mut W

Stop all outputs, execute dead time while fault active

pub fn wait(self) -> &'a mut W

Stop all outputs, jump to next compare cycle and wait

pub fn waitsw(self) -> &'a mut W

Stop all outputs, wait for software action

pub fn edgetrig(self) -> &'a mut W

Stop output on edge, jump to next compare cycle

pub fn edgetrigfreq(self) -> &'a mut W

Stop output on edge, maintain frequency

pub fn lvltrigfreq(self) -> &'a mut W

Stop output at level, maintain frequency